% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lts_extrapolate_numbers.R
\name{lts_extrapolate_numbers}
\alias{lts_extrapolate_numbers}
\title{Title}
\usage{
lts_extrapolate_numbers()
}
\arguments{
\item{weight}{}
}
\value{
a list including weekly estimates of entry and exist
together a visual barchart and  line chart
}
\description{
Description
}
\examples{
lts_extrapolate_numbers()
}
