% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lts_sample_shift.R
\name{lts_sample_shift}
\alias{lts_sample_shift}
\title{lts_sample_shift}
\usage{
lts_sample_shift(
  config_path,
  sample_size = 400,
  start_date,
  end_date,
  strata = "low",
  filename
)
}
\arguments{
\item{config_path}{path to the the configuration file}

\item{sample_size}{sample size neeeded}

\item{start_date}{enforce mdy "11-01-23"}

\item{end_date}{enforce mdy  "11-30-23"}

\item{strata}{"low" "mid-low" "mid-high" "high" if strata is NULL then default is "low"}

\item{filename}{file to save the output}
}
\value{
selection of specific shift selected as sample (day / time / terminal / gate )
}
\description{
Draw the sample of ata collection shift to perform from the full
location time universe
}
\details{
NEXT - if stratified sampling then need to input data collected from previous
shift - this will allow to calculate weight to apply for the next shift
during the extrapolation
}
\examples{
## day1_universe.xlsx
config_path <- system.file("day1_universe.xlsx", package = "LocationTimeSampling")


lts_sample_shift(config_path,
                 sample_size = 400,
                 start_date = "11-01-23", 
                 end_date = "11-30-23",
                 strata = "low",
                 filename = "sampleshift.xlsx")
}
