% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lts_setup.R
\name{lts_setup}
\alias{lts_setup}
\title{lts_setup}
\usage{
lts_setup()
}
\arguments{
\item{location}{(Bus terminal), point (bus gate) and}

\item{time}{(timeshedule for bus).}

\item{constraint}{restriction ad-hic}

\item{filepath}{to save the xlsform}
}
\value{
result will be a list that will include all valid time/location combination within a day
}
\description{
Apply a factorial design on all potential combination of time and location -
includes the possibiity to add ad-hoc constraints
}
\details{
The function will save an xlsform to be used for data collection
preconfigured with the required  data universe
}
\examples{
lts_setup()
}
